+++ fabfile/__init__.py
@@ -1,0 +1,3 @@
+from . import fusion, documint, diamond, tools
+
+__all__ = ['fusion', 'documint', 'diamond', 'tools']
