+++ fabfile/fusion/__init__.py
@@ -1,0 +1,3 @@
+from . import uat
+
+__all__ = ['uat']
